+++ outputs.tf
@@ -33,14 +33,6 @@ output "oidc_issuer_url" {
   value = module.app_aks.oidc_issuer_url
 }
 
-output "storage_account" {
-  value = var.external_bucket != null ? "" : coalesce(var.storage_account, local.account_name, "")
-}
-
-output "storage_container" {
-  value = var.external_bucket != null ? "" : coalesce(var.blob_container, local.container_name)
-}
-
 output "private_link_resource_id" {
   value = module.app_lb.gateway_id
 }
@@ -48,7 +40,7 @@ output "private_link_resource_id" {
 output "private_link_sub_resource_name" {
   value = module.app_lb.frontend_ip_configuration_names
 }
-  
+
 output "standardized_size" {
   value = var.size
 }
@@ -63,4 +55,12 @@ output "aks_node_instance_type" {
 
 output "database_instance_type" {
   value = try(local.deployment_size[var.size].db, var.database_sku_name)
+}
+
+output "client_id" {
+  value = module.identity.identity.client_id
+}
+
+output "tenant_id" {
+  value = module.identity.identity.tenant_id
 }